Sync #75 from devel to testing
ansible: correctly deal with pcs resource when restarting rgw.

The commands 'pcs resource manage/unmanage/delete' only need run once in
the Ceph RadosGW restarting procedure, and the order of them should be
correct.

Before this patch, the steps in the 'restart ceph rgws' playbook are:
1. the first obs node set ceph-rgw pcs resource unmanged, and restart
   ceph-radosgw, and then set ceph-rgw pcs resource managed, finally
   delete the httpd pcs resource;
2. all other obs nodes restart ceph-radosgw one by one.

The correct steps are:
1. set ceph-rgw pcs resource unmanaged;
2. all obs nodes restart ceph-radosgw one by one;
3. set ceph-rgw pcs resource managed;
4. delete httpd pcs resource.

It may be clearer to use a separate playbook for step 1, and another playbook
for step 3 and step 4. But we can use run_once and delegate to achieve
the same procedure.

Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>
(cherry picked from commit 528372ea0a8a0bfe0de8040d0474cd8f335a49c0)
Signed-off-by: Hunt Xu <mhuntxu@gmail.com>