PPTP: persist proxy arp when kicking previous connection
For safety reason, upon the establishment of a new PPTP connection, the
previous connection using the same credential will be terminated. The
proxy arp entry set up by PPTP server will also be removed during the
termination.

The proxy arp entry should be kept. So this commit fixes the problem by
adding the entry back after the termination of the previous connection.

Fixes: redmine #11076

Signed-off-by: Hunt Xu <mhuntxu@gmail.com>